CRL and OCSP test (#17)
CRL test was quite simple, set up a CRL, and then revoke a certificate, then redo the CRL. This CRL file is updated from index.txt and then displays the revocation state of the certificate in questio.
OCSP test was a little more difficult, if you want to set up your own OCSP certificate (still a pem file) you have to create an OCSP cert, as an intermediate certificate, which signs other certificates. Because we do not have the ability to change much within the structure of freeradius we just have to append the details onto the ca config, before the make is run against the freeradius creation of certificates.
This then is overwritten at run-time anyway, and allows us to check. In production, these values are set to "no overwritten OCSP endpoints", which means it will use the correct format of certificates if available.